Changed file location to absolute path and relative path
Added the absolute directory for the script directory and then the
relative path to the file.  In this case the same as the script
directory. This relative path can be amended if you want to put the var
files somewhere else.  NOTE!!!!  the files PeerToLabelMapping,
ServicePrefixes, and bgplog.json … are created by:
getlabelsandserviceprefixes.py in the same directory as the
getlabelsandserviceprefixes python script.